# openapi.yml
openapi: 3.0.3
info:
  title: Pokedex V2 – Backend API
  version: 1.0.0
  description: |
    API REST para gestionar entrenadores, Pokémon y equipos.
    Incluye autenticación con JWT, gestión de equipos y relación Pokémon–Equipo.
servers:
  - url: http://localhost:8080
    description: Servidor local de desarrollo
...
security:
  - bearerAuth: []
paths:
  /api/v2/auth/register:
    post:
      tags: [Auth]
      summary: Registro de entrenador con token JWT
      description: Registra un nuevo entrenador y devuelve los datos del entrenador junto con un token JWT.
      security: []  # público
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntrenadorRequest'
            example:
              pokemon_favorito: "Pikachu"
              nombre: "Ash Ketchum"
              email: "ash@example.com"
              contrasenia: "P@ssw0rd123"
              region_preferida: "Kanto"
              tipo_preferido: "Eléctrico"
      responses:
        '201':
          description: Entrenador registrado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
              example:
                uri: "/api/v2/auth/register"
                message: "Entrenador registrado exitosamente."
                status: 201
                date: "2025-11-20"
                data:
                  entrenador:
                    id_entrenador: 1
                    nombre: "Ash Ketchum"
                    email: "ash@example.com"
                    pokemon_favorito: "Pikachu"
                    region_preferida: "Kanto"
                    tipo_preferido: "Eléctrico"
                  token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  type: "Bearer"
        '400':
          description: Datos de entrada inválidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v2/auth/login:
    post:
      tags: [Auth]
      summary: Login de entrenador (JWT)
      description: Autentica al entrenador por email y contraseña y devuelve un token JWT.
      security: []  # público
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntrenadorLoginRequest'
            example:
              usuario: "ash@example.com"
              password: "P@ssw0rd123"
      responses:
        '200':
          description: Login exitoso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
              example:
                uri: "/api/v2/auth/login"
                message: "Login JWT exitoso."
                status: 200
                date: "2025-11-20"
                data:
                  token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  type: "Bearer"
                  email: "ash@example.com"
        '400':
          description: Credenciales inválidas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: No autorizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v2/entrenador:
    post:
      tags: [Entrenador]
      summary: Crear entrenador (solo pruebas)
      description: Crea un nuevo entrenador (equivalente al registro pero sin token).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntrenadorRequest'
            example:
              pokemon_favorito: "Charmander"
              nombre: "Red"
              email: "red@example.com"
              contrasenia: "Password123"
              region_preferida: "Kanto"
              tipo_preferido: "Fuego"
      responses:
        '201':
          description: Entrenador creado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
        '400':
          description: Datos inválidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    get:
      tags: [Entrenador]
      summary: Obtener todos los entrenadores
      responses:
        '200':
          description: Lista de entrenadores obtenida correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
              example:
                uri: "/api/v2/entrenador"
                message: "Entrenadores encontrados"
                status: 200
                date: "2025-11-20"
                data:
                  - id_entrenador: 1
                    nombre: "Ash"
                    email: "ash@example.com"
                  - id_entrenador: 2
                    nombre: "Misty"
                    email: "misty@example.com"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    patch:
      tags: [Entrenador]
      summary: Actualizar entrenador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntrenadorUpdateRequest'
      responses:
        '200':
          description: Entrenador actualizado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
        '400':
          description: Datos inválidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Entrenador no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v2/entrenador/email/{email}:
    get:
      tags: [Entrenador]
      summary: Obtener entrenador por email
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Entrenador encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
        '404':
          description: Entrenador no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v2/entrenador/region:
    get:
      tags: [Entrenador]
      summary: Obtener entrenadores por región favorita
      parameters:
        - name: region
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Entrenadores encontrados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
  /api/v2/entrenador/tipo:
    get:
      tags: [Entrenador]
      summary: Obtener entrenadores por tipo favorito
      parameters:
        - name: tipo
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Entrenadores encontrados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
  /api/v2/entrenador/pokemon:
    get:
      tags: [Entrenador]
      summary: Obtener entrenadores por Pokémon favorito
      parameters:
        - name: pokemon
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Entrenadores encontrados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
  /api/v2/entrenador/id/{id}:
    delete:
      tags: [Entrenador]
      summary: Eliminar entrenador por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Entrenador eliminado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
        '404':
          description: Entrenador no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v2//entrenador/email/{email}:
    delete:
      tags: [Entrenador]
      summary: Eliminar entrenador por email
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Entrenador eliminado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
        '404':
          description: Entrenador no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /api/v2/pokemon:
    post:
      tags: [Pokemon]
      summary: Crear Pokémon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PokemonRequest'
            example:
              nombre: "Bulbasaur"
              tipo_principal: "Planta"
              tipo_secundario: "Veneno"
              sprite_normal: "https://img.poke/bulbasaur.png"
              sprite_shiny: "https://img.poke/bulbasaur_shiny.png"
      responses:
        '201':
          description: Pokémon creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
    get:
      tags: [Pokemon]
      summary: Obtener todos los Pokémon
      responses:
        '200':
          description: Lista de Pokémon.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'

    patch:
      tags: [Pokemon]
      summary: Actualizar Pokémon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PokemonUpdateRequest'
      responses:
        '200':
          description: Pokémon actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
        '404':
          description: Pokémon no encontrado.
...
  /api/v2/pokemon/{id}:
    delete:
      tags: [Pokemon]
      summary: Eliminar Pokémon por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Pokémon eliminado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralResponse'
        '404':
          description: Pokémon no encontrado.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GeneralResponse:
      type: object
      properties:
        uri:
          type: string
          nullable: true
        message:
          type: string
        status:
          type: integer
          format: int32
        date:
          type: string
          format: date
        data:
          description: Contenido variable de la respuesta (puede ser objeto, lista o valor simple).
      required: [message, status]
    ApiErrorResponse:
      type: object
      properties:
        status:
          type: integer
        date:
          type: string
          format: date
        uri:
          type: string
          nullable: true
        data:
          description: Detalle del error (mensaje, validaciones, etc.)

    # --- Auth / Entrenador ---
    EntrenadorRequest:
      type: object
      properties:
        pokemon_favorito:
          type: string
          nullable: true
        nombre:
          type: string
        email:
          type: string
          format: email
        contrasenia:
          type: string
          minLength: 8
        region_preferida:
          type: string
          nullable: true
        tipo_preferido:
          type: string
          nullable: true
      required:
        - nombre
        - email
        - contrasenia
    EntrenadorLoginRequest:
      type: object
      properties:
        usuario:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - usuario
        - password
    EntrenadorUpdateRequest:
      type: object
      description: |
  
      properties:
        id_entrenador:
          type: integer
          format: int64
        nombre:
          type: string
        pokemon_favorito:
          type: string
        region_preferida:
          type: string
        tipo_preferido:
          type: string
...
    PokemonRequest:
      type: object
      properties:
        nombre:
          type: string
        tipo_principal:
          type: string
        tipo_secundario:
          type: string
          nullable: true
        sprite_normal:
          type: string
          nullable: true
        sprite_shiny:
          type: string
          nullable: true
      required:
        - nombre
        - tipo_principal
    PokemonUpdateRequest:
      type: object
      properties:
        id_pokemon:
          type: integer
          format: int64
        tipo_principal:
          type: string
        tipo_secundario:
          type: string
          nullable: true
        sprite_normal:
          type: string
          nullable: true
        sprite_shiny:
          type: string
          nullable: true
      required:
        - id_pokemon
...